[Clean & update] Clean project and update dependencies
- move kotlin source files from kotlin to java folder so we don't need to care about "sourceSets"
- remove "sourceSets" in gradle files because it's not needed
- remove files that are in gitignore (e.g. gradle.properties)
