--- .idea/modules.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/SharedSequence/SharedSequence.iml" filepath="$PROJECT_DIR$/SharedSequence/SharedSequence.iml" />
-      <module fileurl="file://$PROJECT_DIR$/SharedSequence.kt.iml" filepath="$PROJECT_DIR$/SharedSequence.kt.iml" />
-      <module fileurl="file://$PROJECT_DIR$/SharedSequenceApi/SharedSequenceApi.iml" filepath="$PROJECT_DIR$/SharedSequenceApi/SharedSequenceApi.iml" />
-      <module fileurl="file://$PROJECT_DIR$/SharedSequenceProcessor/SharedSequenceProcessor.iml" filepath="$PROJECT_DIR$/SharedSequenceProcessor/SharedSequenceProcessor.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-    </modules>
-  </component>
-</project>
