+++ SharedSequence/src/main/java/org/notests/sharedsequence/DriverTraits.kt
