+++ SharedSequence/src/main/java/org/notests/sharedsequence/SignalTraits.kt
