+++ SharedSequence/src/test/java/org/notests/sharedsequence/utils/MyTestSubscriber.kt
