+++ SharedSequenceApi/build.gradle
@@ -10,9 +10,5 @@ dependencies {
     implementation 'io.reactivex.rxjava2:rxjava:' + versions.rxjava2
 }
 
-sourceSets {
-    main.java.srcDirs += "src/main/kotlin"
-}
-
 sourceCompatibility = "1.7"
 targetCompatibility = "1.7"
