+++ SharedSequenceApi/src/main/java/org/notests/sharedsequence/api/ErrorReporting.kt
