+++ SharedSequenceApi/src/main/java/org/notests/sharedsequence/api/SharedSequence.kt
