+++ SharedSequenceApi/src/main/java/org/notests/sharedsequence/api/SharingTraits.kt
