+++ SharedSequenceProcessor/build.gradle
@@ -14,10 +14,6 @@ dependencies {
 sourceCompatibility = "1.7"
 targetCompatibility = "1.7"
 
-sourceSets {
-    main.java.srcDirs += "src/main/kotlin"
-}
-
 buildscript {
     repositories {
         mavenCentral()
