+++ build.gradle
@@ -7,7 +7,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:3.3.1'
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:' + versions.kotlin
 
         // NOTE: Do not place your application dependencies here; they belong
