+++ versions.gradle
@@ -5,9 +5,9 @@
 ext.versions = [:]
 
 def versions = [:]
-versions.rxjava2 = '2.2.0'
-versions.rx_android = '2.0.2'
-versions.kotlin = '1.2.51'
+versions.rxjava2 = '2.2.7'
+versions.rx_android = '2.1.1'
+versions.kotlin = '1.3.21'
 versions.support = '26.1.0'
 versions.constraint_layout = '1.1.2'
 versions.rxbinding = '2.1.1'
